Add moar puppetlabs modules
This commit adds more puppetlabs "supported" modules, particularly in
the cloud providers section. Special mention goes to azure_arm which
manages to beat dsc on the chonk factor with over 8k files and kuberbetes
which adds a respectiible 1k in files. Dsc is switched over
to using a git source as the file count doubles.

Using a git source mirrors what many customers do by cloning modules
and pushing them to internal git servers. Also, we're trying to really
kick Code Manger in the the teeth, so we must show no mercy.